requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings